+++ src/main/java/ssu/eatssu/domain/partnership/dto/request/CreatePartnershipRequest.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.partnership.dto;
+package ssu.eatssu.domain.partnership.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
