+++ src/main/java/ssu/eatssu/domain/partnership/presentation/PartnershipController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ssu.eatssu.domain.auth.security.CustomUserDetails;
-import ssu.eatssu.domain.partnership.dto.CreatePartnershipRequest;
-import ssu.eatssu.domain.partnership.dto.PartnershipResponse;
+import ssu.eatssu.domain.partnership.dto.request.CreatePartnershipRequest;
+import ssu.eatssu.domain.partnership.dto.response.PartnershipResponse;
 import ssu.eatssu.domain.partnership.presentation.docs.PartnershipControllerDocs;
 import ssu.eatssu.domain.partnership.service.PartnershipService;
 import ssu.eatssu.global.handler.response.BaseResponse;
